+++ packages/orq-rc/RELEASES.md
@@ -8189,3 +8189,13 @@ Based on:
 - [typescript v4.8.0-rc.3] packages/orq-rc
 ### Releases
 - [NPM v4.8.0-rc.3] https://www.npmjs.com/package/@orq-ai/node/v/4.8.0-rc.3 - packages/orq-rc
+
+## 2026-04-10 14:15:53
+### Changes
+Based on:
+- OpenAPI Doc  
+- Speakeasy CLI 1.761.2 (2.879.10) https://github.com/speakeasy-api/speakeasy
+### Generated
+- [typescript v4.8.0-rc.4] packages/orq-rc
+### Releases
+- [NPM v4.8.0-rc.4] https://www.npmjs.com/package/@orq-ai/node/v/4.8.0-rc.4 - packages/orq-rc